Add node example picker
