+++ package.json
@@ -10,6 +10,8 @@
   },
   "devDependencies": {
     "@colony/eslint-config-colony": "^9.0.2",
+    "@types/inquirer": "^8.2.1",
+    "@types/yargs": "^17.0.10",
     "@typescript-eslint/eslint-plugin": "^5.17.0",
     "@typescript-eslint/parser": "^5.17.0",
     "esbuild": "^0.14.27",
@@ -20,7 +22,9 @@
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-prettier": "^4.0.0",
     "ethers": "^5.6.1",
+    "fast-glob": "^3.2.11",
     "husky": "^7.0.4",
+    "inquirer": "^8.2.4",
     "lint-staged": "^12.4.0",
     "prettier": "^2.6.1",
     "release-it": "^14.14.2",
@@ -28,13 +32,14 @@
     "ts-node": "^10.7.0",
     "typedoc": "^0.22.15",
     "typedoc-plugin-markdown": "^3.12.0",
-    "typescript": "^4.6.2"
+    "typescript": "^4.6.2",
+    "yargs": "^17.5.1"
   },
   "peerDependencies": {
     "ethers": "^5.6.1"
   },
   "scripts": {
-    "example:node": "npm run compile-cjs && ts-node src/index.ts",
+    "example:node": "ts-node examples/node/index.ts",
     "example:browser": "esbuild --bundle examples/browser/src/*.ts --servedir=examples/browser/web",
     "build": "npm run clean && npm run compile-cjs && npm run compile-esm && npm run compile-types && npm run build-docs",
     "build-docs": "typedoc",
